Revert "account header syscalls handlers initialized"
This reverts commit cb585ec63475c7812eb80bf80f1ff7bbd0f88970.
